requirements.txt
azure-core==1.30.1
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
fastapi==0.110.1
openai==1.14.2
pydantic==2.6.4
pydantic_core==2.16.3
uvicorn==0.29.0
argparse==1.4.0